refactor(redis_ops): extract connection module
Extract Redis connection and key discovery functions into connection.rs:
- connect_redis: Establishes async Redis connection with optional auth
- scan_keys: Scans Redis keys using SCAN command with pattern matching
