+++ src/redis_ops/mod.rs
@@ -1,6 +1,8 @@
+mod connection;
 mod formatting;
 mod protocol;
 
+pub(crate) use connection::{connect_redis, scan_keys};
 pub(crate) use formatting::{append_expire_command, quoted};
 pub(crate) use protocol::format_command_output;
 
@@ -17,58 +19,6 @@ use crate::{
     config::{DumpConfig, OutputFormat},
     progress::DumpProgress,
 };
-
-async fn connect_redis(config: &DumpConfig) -> Result<redis::aio::MultiplexedConnection> {
-    let redis_url = if let Some(password) = &config.password {
-        format!(
-            "redis://{}:{}@{}:{}",
-            config.user, password, config.host, config.port
-        )
-    } else {
-        format!("redis://{}:{}", config.host, config.port)
-    };
-
-    let client = redis::Client::open(redis_url)?;
-    let mut connection = client.get_multiplexed_async_connection().await?;
-
-    if let Some(db) = config.database {
-        let _: () = redis::cmd("SELECT")
-            .arg(db)
-            .query_async(&mut connection)
-            .await?;
-    }
-
-    Ok(connection)
-}
-
-async fn scan_keys(
-    connection: &mut redis::aio::MultiplexedConnection,
-    pattern: &str,
-    scan_size: usize,
-) -> Result<Vec<String>> {
-    let mut cursor = 0;
-    let mut all_keys = Vec::new();
-
-    loop {
-        let (new_cursor, keys): (u64, Vec<String>) = redis::cmd("SCAN")
-            .arg(cursor)
-            .arg("MATCH")
-            .arg(pattern)
-            .arg("COUNT")
-            .arg(scan_size)
-            .query_async(connection)
-            .await?;
-
-        all_keys.extend(keys);
-        cursor = new_cursor;
-
-        if cursor == 0 {
-            break;
-        }
-    }
-
-    Ok(all_keys)
-}
 
 // --- Key Metadata ---
